requirements.txt
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Gravatar==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
pip==22.0.3
wheel==0.37.1
python-decouple==3.6
